Remade the whole system in preparation for rendering a glyphs in a 'fancy' way
Separate glyphs (aka characters) are now consist of symbols (previously glyphs),
and symbols do consist of 4 parts each. Part can be either a line or a none (for now).
Separated parsing and rendering of a text into different classes. Parsing a text means converting it into array of words, which consist of glyphs and glyphs you know already

what a mess